+++ WinFormAnimation/Path2DExtensions.cs
@@ -41,7 +41,7 @@ namespace WinFormAnimation
         /// <param name="duration">Duration of the animation</param>
         /// <param name="delay">Starting delay</param>
         /// <returns>An array of paths including the newly created one</returns>
-        public static Path2D[] ContinueTo(this Path2D[] paths, Float2D end, ulong duration, float delay)
+        public static Path2D[] ContinueTo(this Path2D[] paths, Float2D end, ulong duration, ulong delay)
         {
             return paths.Concat(new[] {new Path2D(paths.Last().End, end, duration, delay)}).ToArray();
         }
@@ -55,7 +55,7 @@ namespace WinFormAnimation
         /// <param name="delay">Starting delay</param>
         /// <param name="function">Animation controller function</param>
         /// <returns>An array of paths including the newly created one</returns>
-        public static Path2D[] ContinueTo(this Path2D[] paths, Float2D end, ulong duration, float delay,
+        public static Path2D[] ContinueTo(this Path2D[] paths, Float2D end, ulong duration, ulong delay,
             AnimationFunctions.Function function)
         {
             return paths.Concat(new[] {new Path2D(paths.Last().End, end, duration, delay, function)}).ToArray();
@@ -100,7 +100,7 @@ namespace WinFormAnimation
         /// <param name="duration">Duration of the animation</param>
         /// <param name="delay">Starting delay</param>
         /// <returns>An array of paths including the newly created one</returns>
-        public static Path2D[] ContinueTo(this Path2D[] paths, float endX, float endY, ulong duration, float delay)
+        public static Path2D[] ContinueTo(this Path2D[] paths, float endX, float endY, ulong duration, ulong delay)
         {
             return
                 paths.Concat(new[] {new Path2D(paths.Last().End, new Float2D(endX, endY), duration, delay)}).ToArray();
@@ -116,7 +116,7 @@ namespace WinFormAnimation
         /// <param name="delay">Starting delay</param>
         /// <param name="function">Animation controller function</param>
         /// <returns>An array of paths including the newly created one</returns>
-        public static Path2D[] ContinueTo(this Path2D[] paths, float endX, float endY, ulong duration, float delay,
+        public static Path2D[] ContinueTo(this Path2D[] paths, float endX, float endY, ulong duration, ulong delay,
             AnimationFunctions.Function function)
         {
             return
@@ -158,7 +158,7 @@ namespace WinFormAnimation
         /// <param name="duration">Duration of the animation</param>
         /// <param name="delay">Starting delay</param>
         /// <returns>An array of paths including the newly created one</returns>
-        public static Path2D[] ContinueTo(this Path2D path, Float2D end, ulong duration, float delay)
+        public static Path2D[] ContinueTo(this Path2D path, Float2D end, ulong duration, ulong delay)
         {
             return path.ToArray().ContinueTo(end, duration, delay);
         }
@@ -172,7 +172,7 @@ namespace WinFormAnimation
         /// <param name="delay">Starting delay</param>
         /// <param name="function">Animation controller function</param>
         /// <returns>An array of paths including the newly created one</returns>
-        public static Path2D[] ContinueTo(this Path2D path, Float2D end, ulong duration, float delay,
+        public static Path2D[] ContinueTo(this Path2D path, Float2D end, ulong duration, ulong delay,
             AnimationFunctions.Function function)
         {
             return path.ToArray().ContinueTo(end, duration, delay, function);
@@ -215,7 +215,7 @@ namespace WinFormAnimation
         /// <param name="duration">Duration of the animation</param>
         /// <param name="delay">Starting delay</param>
         /// <returns>An array of paths including the newly created one</returns>
-        public static Path2D[] ContinueTo(this Path2D path, float endX, float endY, ulong duration, float delay)
+        public static Path2D[] ContinueTo(this Path2D path, float endX, float endY, ulong duration, ulong delay)
         {
             return path.ToArray().ContinueTo(endX, endY, duration, delay);
         }
@@ -230,7 +230,7 @@ namespace WinFormAnimation
         /// <param name="delay">Starting delay</param>
         /// <param name="function">Animation controller function</param>
         /// <returns>An array of paths including the newly created one</returns>
-        public static Path2D[] ContinueTo(this Path2D path, float endX, float endY, ulong duration, float delay,
+        public static Path2D[] ContinueTo(this Path2D path, float endX, float endY, ulong duration, ulong delay,
             AnimationFunctions.Function function)
         {
             return path.ToArray().ContinueTo(endX, endY, duration, delay, function);
